// src/openapi/client/models/AdvertisementWithFiltersResponse.ts
export interface AdvertisementWithFiltersResponse {
    /**
     * 
     * @type {Array<AdvertisementFilterParameterResponse>}
     * @memberof AdvertisementWithFiltersResponse
     */
    filterParameters?: Array<AdvertisementFilterParameterResponse>;
    /**
     * 
     * @type {number}
     * @memberof AdvertisementWithFiltersResponse
     */
    id: number;
    /**
     * 
     * @type {string}
     * @memberof AdvertisementWithFiltersResponse
     */
    brand: string;
    /**
     * 
     * @type {string}
     * @memberof AdvertisementWithFiltersResponse
     */
    model: string;
    /**
     * 
     * @type {string}
     * @memberof AdvertisementWithFiltersResponse
     */
    price: string;
    /**
     * 
     * @type {string}
     * @memberof AdvertisementWithFiltersResponse
     */
    currency: string;
    /**
     * 
     * @type {number}
     * @memberof AdvertisementWithFiltersResponse
     */
    releaseYear: number;
    /**
     * 
     * @type {string}
     * @memberof AdvertisementWithFiltersResponse
     */
    region: string;
    /**
     * 
     * @type {Date}
     * @memberof AdvertisementWithFiltersResponse
     */
    createdAt: Date;
    /**
     * 
     * @type {Array<string>}
     * @memberof AdvertisementWithFiltersResponse
     */
    images?: Array<string>;
}